Split wsgi tests into multiple files
